pyproject.toml
[project]
name = "airflow-pyspark-docker"
version = "0.1.0"
description = "Minimal Project Setup for our future ETL pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy==2.2.5",
    "pandas==2.2.3",
    "polars==1.29.0",
    "py4j==0.10.9.7",
    "pyspark==3.5.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "six==1.17.0",
    "tzdata==2025.2"
]
